refactor: modularize letter generation logic
